neuron_minimal_io/pyproject.toml
[tool.poetry]
name = "neuron-minimal-io"
version = "0.1.0"
description = ""
authors = ["Massimiliano della Rovere <massimiliano.dellarovere@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyserial = "^3.5"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
